requirements.txt
click==6.7
dominate==2.3.1
Flask==0.12.2
flask-appconfig==0.11.1
Flask-Bootstrap==3.3.7.1
Flask-Debug==0.4.3
flask-nav==0.6
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
#pip==9.0.1
#PyMySQL==0.7.11
MySQL-python
robotframework==3.0.2
#setuptools==36.0.1
six==1.10.0
SQLAlchemy==1.1.10
visitor==0.1.3
Werkzeug==0.12.2
#wheel==0.29.0
WTForms==2.1
appdirs==1.4.3
asn1crypto==0.22.0
boto3==1.4.4
botocore==1.5.58
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
cryptography==1.9
decorator==4.0.11
docutils==0.13.1
enum34==1.1.6
futures==3.1.1
idna==2.5
ipaddress==1.0.18
jmespath==0.9.3
packaging==16.8
paramiko==2.1.2
pyasn1==0.2.3
pycparser==2.17
pyparsing==2.2.0
python-dateutil==2.6.0
PyYAML==3.12
requests==2.17.3
robotframework-selenium2library==1.8.0
s3transfer==0.1.10
selenium==3.4.3
six==1.10.0
urllib3==1.21.1
xlrd==1.0.0
robotframework-sshlibrary==2.1.3
